refactor: #23 prep
