+++ .golangci.yaml
@@ -7,18 +7,36 @@ run:
   timeout: 180s
 
 linters:
-  disable-all: false
   enable-all: true
   disable:
     - tenv
     - godox
     - gci
     - gofumpt
-    - godot
     - gofmt
     - wsl
+    - typecheck
 
 linters-settings:
+  unused:
+    # Mark all struct fields that have been written to as used.
+    # Default: true
+    field-writes-are-uses: false
+    # Treat IncDec statement (e.g. `i++` or `i--`) as both read and write operation instead of just write.
+    # Default: false
+    post-statements-are-reads: true
+    # Mark all exported fields as used.
+    # default: true
+    exported-fields-are-used: false
+    # Mark all function parameters as used.
+    # default: true
+    parameters-are-used: false
+    # Mark all local variables as used.
+    # default: true
+    local-variables-are-used: false
+    # Mark all identifiers inside generated files as used.
+    # Default: true
+    generated-is-used: false
   depguard:
     # Rules to apply.
     #
