--- src/functions/core/process.go
@@ -1,130 +1,0 @@
-package core
-
-import (
-	"Scruticode/src/constants"
-	"Scruticode/src/functions/options"
-	"Scruticode/src/functions/utils"
-	"Scruticode/src/functions/validations"
-	"Scruticode/src/types"
-	"log"
-	"strings"
-)
-
-func ProcessConfigFile(content string) {
-	sections := ExtractSections(content)
-
-	for _, section := range sections {
-		_, keyValues := ParseSection(section)
-		// Check if is needed here to validate the HEADER, right now it's being ignored
-		processKeyValues(keyValues)
-	}
-}
-
-func ExtractSections(content string) []string {
-	return strings.Split(content, "\n\n")
-}
-
-func ParseSection(section string) (string, []string) {
-	lines := strings.Split(section, "\n")
-	if len(lines) == 0 {
-		return "", nil
-	}
-
-	var keyValues []string
-	var header string
-
-	for index, line := range lines {
-		trimmedLine := strings.TrimSpace(line)
-		if trimmedLine == "" || IsComment(trimmedLine) {
-			continue
-		}
-
-		if index == constants.IsEmpty && IsSectionHeader(trimmedLine) {
-			header = trimmedLine
-
-			continue
-		}
-
-		if strings.Contains(trimmedLine, "=") {
-			keyValues = append(keyValues, trimmedLine)
-		}
-	}
-
-	return header, keyValues
-}
-
-func processKeyValues(keyValues []string) {
-	var actions = map[string]types.ActionFunc{
-		"docker_compose":       func() { log.Println("action for docker_compose") },
-		"dockerfile":           func() { log.Print(options.DockerfileExists()) },
-		"readme":               func() { log.Print(options.Readme()) },
-		"ci":                   func() { log.Println("action for ci") },
-		"cd":                   func() { log.Println("action for cd") },
-		"conventional_commits": func() { log.Println("action for conventional_commits") },
-		"pre_commit":           func() { log.Println("action for pre_commit") },
-		"linter":               func() { log.Println("action for linter") },
-		"formatter":            func() { log.Println("action for formatter") },
-		"unit":                 func() { log.Println("action for unit") },
-		"integration":          func() { log.Println("action for integration") },
-		"e2e":                  func() { log.Println("action for e2e") },
-		"coverage":             func() { log.Println("action for coverage") },
-		"stress":               func() { log.Println("action for stress") },
-		"secrets":              func() { log.Println("action for secrets") },
-		"iac":                  func() { log.Println("action for iac") },
-		"code":                 func() { log.Println("action for code") },
-		"container":            func() { log.Println("action for container") },
-		"deps":                 func() { log.Println("action for deps") },
-		"sast":                 func() { log.Println("action for sast") },
-		"dast":                 func() { log.Println("action for dast") },
-	}
-
-	const emptyAsString = ""
-	for _, pair := range keyValues {
-		key, value := ParseKeyValuePair(pair)
-		if key == emptyAsString {
-			continue
-		}
-
-		keyAsString := utils.ToAbsoluteString(value)
-		keyLangOrPlatform := key == "langs" || key == "platforms"
-
-		if keyLangOrPlatform {
-			validations.ExtraLangConfig(keyAsString)
-			validations.ExtraPlatformConfig(keyAsString)
-		}
-
-		if IsActionEnabled(value) {
-			action, exists := actions[key]
-			if !exists {
-				log.Printf("No action found for key '%s'\n", key)
-
-				return
-			}
-			action()
-		}
-	}
-}
-
-func IsSectionHeader(line string) bool {
-	return strings.HasPrefix(line, "[") && strings.HasSuffix(line, "]")
-}
-
-func IsComment(line string) bool {
-	return strings.HasPrefix(strings.TrimSpace(line), "#")
-}
-
-func IsActionEnabled(action string) bool {
-	return strings.ToLower(action) == "true"
-}
-
-func ParseKeyValuePair(pair string) (string, string) {
-	parts := strings.SplitN(pair, "=", constants.IsKeyVal)
-	if len(parts) != constants.IsKeyVal {
-		return "", ""
-	}
-
-	key := strings.TrimSpace(parts[constants.IsFirstIndex])
-	value := strings.TrimSpace(parts[constants.IsSecondIndex])
-
-	return key, value
-}
