+++ src/main.go
@@ -1,13 +1,9 @@
 package main
 
-import (
-	"Scruticode/src/config"
-	"Scruticode/src/functions/arguments"
-	"Scruticode/src/functions/core"
-)
+import "Scruticode/src/core/functions"
 
 func main() {
-	configuration := config.ReadConfigFile()
-	core.ProcessConfigFile(configuration)
-	arguments.Generate()
+	configuration := functions.ReadConfigFile()
+	functions.ProcessConfigFile(configuration)
+	functions.Generate()
 }
