+++ src/shared/constants/errors.go
