--- src/types/process.go
@@ -1,3 +1,0 @@
-package types
-
-type ActionFunc func()
